DRTII-972 Change pax info fetching mechanism  - stop fetching all flights data for 3 whole days in one go  - lazy load individual flights when graph icon clicked
