+++ client/src/main/scala/drt/client/components/FlightsWithSplitsTable.scala
@@ -34,7 +34,6 @@ import scala.collection.immutable.{HashSet, Map}
 
 object FlightsWithSplitsTable {
   case class Props(flightsWithSplits: List[ApiFlightWithSplits],
-                   passengerInfoSummaryByDay: Map[UtcDate, Map[ArrivalKey, PassengerInfoSummary]],
                    queueOrder: Seq[Queue], hasEstChox: Boolean,
                    arrivalSources: Option[(UniqueArrival, Pot[List[Option[FeedSourceArrival]]])],
                    loggedInUser: LoggedInUser,
@@ -49,6 +48,8 @@ object FlightsWithSplitsTable {
                    redListUpdates: RedListUpdates,
                    airportConfig: AirportConfig,
                   ) extends UseValueEq
+
+  case class State(apiDataLoaded: Boolean)
 
   def ArrivalsTable(timelineComponent: Option[Arrival => VdomNode] = None,
                     originMapper: PortCode => VdomNode = portCode => portCode.toString,
@@ -94,16 +95,11 @@ object FlightsWithSplitsTable {
             <.tbody(
               sortedFlights.zipWithIndex.map {
                 case ((flightWithSplits, codeShares), idx) =>
-                  val maybePassengerInfo: Option[PassengerInfoSummary] = props
-                    .passengerInfoSummaryByDay
-                    .get(SDate(flightWithSplits.apiFlight.Scheduled).toUtcDate)
-                    .flatMap(_.get(ArrivalKey(flightWithSplits.apiFlight)))
                   val isRedListOrigin = props.redListPorts.contains(flightWithSplits.apiFlight.Origin)
                   val directRedListFlight = redlist.DirectRedListFlight(props.viewMode.dayEnd.millisSinceEpoch, props.portCode, props.terminal, flightWithSplits.apiFlight.Terminal, isRedListOrigin)
                   val redListPaxInfo = redlist.IndirectRedListPax(props.displayRedListInfo, flightWithSplits)
                   FlightTableRow.component(FlightTableRow.Props(
                     flightWithSplits = flightWithSplits,
-                    maybePassengerInfoSummary = maybePassengerInfo,
                     codeShares = codeShares,
                     idx = idx,
                     timelineComponent = timelineComponent,
@@ -208,202 +204,3 @@ object FlightsWithSplitsTable {
   }
 }
 
-object FlightTableRow {
-
-  import FlightTableComponents._
-
-  type OriginMapperF = PortCode => VdomNode
-
-  type SplitsGraphComponentFn = SplitsGraph.Props => TagOf[Div]
-
-  case class Props(flightWithSplits: ApiFlightWithSplits,
-                   maybePassengerInfoSummary: Option[PassengerInfoSummary],
-                   codeShares: Set[Arrival],
-                   idx: Int,
-                   timelineComponent: Option[Arrival => html_<^.VdomNode],
-                   originMapper: OriginMapperF = portCode => portCode.toString,
-                   splitsGraphComponent: SplitsGraphComponentFn = (_: SplitsGraph.Props) => <.div(),
-                   splitsQueueOrder: Seq[Queue],
-                   hasEstChox: Boolean,
-                   loggedInUser: LoggedInUser,
-                   viewMode: ViewMode,
-                   defaultWalkTime: Long,
-                   hasTransfer: Boolean,
-                   indirectRedListPax: IndirectRedListPax,
-                   directRedListFlight: DirectRedListFlight,
-                   airportConfig: AirportConfig,
-                   redListUpdates: RedListUpdates,
-                  ) extends UseValueEq
-
-  case class RowState(hasChanged: Boolean)
-
-  val component: Component[Props, RowState, Unit, CtorType.Props] = ScalaComponent.builder[Props](displayName = "TableRow")
-    .initialState[RowState](RowState(false))
-    .render_PS((props, state) => {
-      val codeShares = props.codeShares
-      val flightWithSplits = props.flightWithSplits
-      val flight = flightWithSplits.apiFlight
-      val allCodes = flight.flightCodeString :: codeShares.map(_.flightCodeString).toList
-
-      val hasChangedStyle = if (state.hasChanged) ^.background := "rgba(255, 200, 200, 0.5) " else ^.outline := ""
-      val timeIndicatorClass = if (flight.PcpTime.getOrElse(0L) < SDate.now().millisSinceEpoch) "before-now" else "from-now"
-
-      val queuePax: Map[Queue, Int] = ApiSplitsToSplitRatio
-        .paxPerQueueUsingBestSplitsAsRatio(flightWithSplits).getOrElse(Map[Queue, Int]())
-
-      val flightCodeClass = if (props.loggedInUser.hasRole(ArrivalSource))
-        "arrivals__table__flight-code arrivals__table__flight-code--clickable"
-      else
-        "arrivals__table__flight-code"
-
-      def flightCodeElement(flightCodes: String, outgoingDiversion: Boolean, incomingDiversion: Boolean): VdomTagOf[Span] =
-        if (props.loggedInUser.hasRole(ArrivalSource)) {
-          val diversionClass = (outgoingDiversion, incomingDiversion) match {
-            case (_, true) => "arrivals__table__flight-code-incoming-diversion"
-            case (true, _) => "arrivals__table__flight-code-outgoing-diversion"
-            case _ => ""
-          }
-          <.span(
-            ^.cls := s"arrivals__table__flight-code-value $diversionClass",
-            ^.onClick --> Callback(SPACircuit.dispatch {
-              props.viewMode match {
-                case vm: ViewDay if vm.isHistoric(SDate.now()) =>
-                  GetArrivalSourcesForPointInTime(props.viewMode.time.addHours(28), props.flightWithSplits.unique)
-                case _: ViewPointInTime =>
-                  GetArrivalSourcesForPointInTime(props.viewMode.time, props.flightWithSplits.unique)
-                case _ =>
-                  GetArrivalSources(props.flightWithSplits.unique)
-              }
-            }),
-            flightCodes,
-          )
-        } else
-          <.span(
-            ^.cls := "arrivals__table__flight-code-value",
-            flightCodes
-          )
-
-      val outgoingDiversion = props.directRedListFlight.outgoingDiversion
-      val ctaOrRedListMarker = if (flight.Origin.isDomesticOrCta) "*" else ""
-      val flightCodes = s"${allCodes.mkString(" - ")}$ctaOrRedListMarker"
-
-      val firstCells = List[TagMod](
-        <.td(^.className := flightCodeClass,
-          <.div(
-            ^.cls := "arrivals__table__flight-code-wrapper",
-            flightCodeElement(flightCodes, props.directRedListFlight.outgoingDiversion, props.directRedListFlight.incomingDiversion),
-            if (flightWithSplits.hasValidApi)
-              props
-                .maybePassengerInfoSummary
-                .map { info =>
-                  <.div(^.className := "arrivals__table__flight-code__info",
-                    FlightChartComponent(FlightChartComponent.Props(flightWithSplits, info)))
-                }
-            else EmptyVdom
-          )),
-        <.td(props.originMapper(flight.Origin)),
-        <.td(TerminalContentComponent.airportWrapper(flight.Origin) { proxy: ModelProxy[Pot[AirportInfo]] =>
-          <.span(
-            proxy().renderEmpty(<.span()),
-            proxy().render(ai => {
-              val style = if (props.indirectRedListPax.isEnabled && isRedListCountry(ai.country, props.viewMode.dayEnd, props.redListUpdates)) {
-                ScalaCssReact.scalacssStyleaToTagMod(ArrivalsPageStylesDefault.redListCountryField)
-              } else EmptyVdom
-
-              <.span(
-                style,
-                ai.country
-              )
-            })
-          )
-        }),
-        props.indirectRedListPax match {
-          case NoIndirectRedListPax => EmptyVdom
-          case NeboIndirectRedListPax(Some(pax)) => <.td(<.span(^.className := "badge", pax))
-          case NeboIndirectRedListPax(None) => <.td(EmptyVdom)
-        },
-        <.td(gateOrStand(flight, props.airportConfig, props.directRedListFlight.paxDiversion)),
-        <.td(flight.displayStatus.description),
-        <.td(localDateTimeWithPopup(Option(flight.Scheduled))),
-        if (props.airportConfig.useTimePredictions) <.td(localDateTimeWithPopup(flight.PredictedTouchdown)) else EmptyVdom,
-        <.td(localDateTimeWithPopup(flight.Estimated)),
-        <.td(localDateTimeWithPopup(flight.Actual)),
-      )
-      val estCell = List(<.td(localDateTimeWithPopup(flight.EstimatedChox)))
-      val lastCells = List[TagMod](
-        <.td(localDateTimeWithPopup(flight.ActualChox)),
-        <.td(pcpTimeRange(flightWithSplits), ^.className := "arrivals__table__flight-est-pcp"),
-        <.td(FlightComponents.paxComp(flightWithSplits, props.directRedListFlight, flight.Origin.isDomesticOrCta))
-      )
-      val flightFields = if (props.hasEstChox) firstCells ++ estCell ++ lastCells else firstCells ++ lastCells
-
-      val paxClass = FlightComponents.paxClassFromSplits(flightWithSplits)
-
-      val flightId = flight.uniqueId.toString
-
-      val timeLineTagMod = props.timelineComponent.map(timeline => <.td(timeline(flight))).toList.toTagMod
-
-      val cancelledClass = if (flight.isCancelled) " arrival-cancelled" else ""
-      val noPcpPax = if (flight.Origin.isCta || outgoingDiversion) " arrival-cta" else ""
-      val trClassName = s"${offScheduleClass(flight, props.airportConfig.timeToChoxMillis, props.airportConfig.useTimePredictions)} $timeIndicatorClass$cancelledClass$noPcpPax"
-
-      val queueTagMod = props.splitsQueueOrder.map { q =>
-        val pax = if (!flight.Origin.isDomesticOrCta) queuePax.getOrElse(q, 0).toString else "-"
-        <.td(^.className := s"queue-split $paxClass right",
-          <.span(pax, ^.className := s"${q.toString.toLowerCase()}-queue-pax"))
-      }.toTagMod
-
-      if (props.hasTransfer) {
-        <.tr(
-          ^.key := flightId,
-          ^.className := trClassName,
-          hasChangedStyle,
-          timeLineTagMod,
-          flightFields.toTagMod,
-          queueTagMod,
-          <.td(FlightComponents.paxTransferComponent(flight))
-        )
-      } else {
-        <.tr(
-          ^.key := flightId,
-          ^.className := trClassName,
-          hasChangedStyle,
-          timeLineTagMod,
-          flightFields.toTagMod,
-          queueTagMod
-        )
-      }
-
-    })
-    .build
-
-  private def gateOrStand(arrival: Arrival, airportConfig: AirportConfig, paxAreDiverted: Boolean): VdomTagOf[Span] = {
-    val gateOrStand = <.span(s"${arrival.Gate.getOrElse("")} / ${arrival.Stand.getOrElse("")}")
-    arrival.walkTime(airportConfig.timeToChoxMillis, airportConfig.firstPaxOffMillis, airportConfig.useTimePredictions).map { wt =>
-      val description = (paxAreDiverted, arrival.Stand.isDefined, arrival.Gate.isDefined) match {
-        case (true, _, _) => "walk time including transfer bus"
-        case (false, true, _) => "walk time from stand"
-        case (false, false, true) => "walk time from gate"
-        case _ => "default walk time"
-      }
-      val walkTimeString = MinuteAsAdjective(millisToMinutes(wt)).display + " " + description
-      <.span(Tippy.interactive(<.span(walkTimeString), gateOrStand))
-    }.getOrElse(gateOrStand)
-  }
-
-  def offScheduleClass(arrival: Arrival, timeToChoxMillis: Long, considerPredictions: Boolean): String = {
-    val eta = arrival.bestArrivalTime(timeToChoxMillis, considerPredictions)
-    val differenceFromScheduled = eta - arrival.Scheduled
-    val hourInMillis = 3600000
-    val offScheduleClass = if (differenceFromScheduled > hourInMillis || differenceFromScheduled < -1 * hourInMillis)
-      "danger"
-    else ""
-    offScheduleClass
-  }
-
-  def isRedListCountry(country: String, date: SDateLike, redListUpdates: RedListUpdates): Boolean =
-    redListUpdates.countryCodesByName(date.millisSinceEpoch).keys.exists(_.toLowerCase == country.toLowerCase)
-
-  def apply(props: Props): Unmounted[Props, RowState, Unit] = component(props)
-
-}
